get offers from store in every component; move cities to constants; filter offers by city; empty offers list component;
