+++ src/store/actions.ts
@@ -1,5 +1,4 @@
 import {createAction} from '@reduxjs/toolkit';
-import {City, OfferListItem} from '@/types';
+import {City} from '@/types';
 
 export const setCurrentCity = createAction<City>('offer/setCurrentCity');
-export const setOffers = createAction<OfferListItem[]>('offer/setList');
